+++ client/v2/autocli/common.go
@@ -9,10 +9,10 @@ import (
 	"sigs.k8s.io/yaml"
 
 	autocliv1 "cosmossdk.io/api/cosmos/autocli/v1"
-	"cosmossdk.io/client/v2/internal/flags"
-	"cosmossdk.io/client/v2/internal/util"
 
 	"github.com/cosmos/cosmos-sdk/client"
+	"github.com/cosmos/cosmos-sdk/client/v2/internal/flags"
+	"github.com/cosmos/cosmos-sdk/client/v2/internal/util"
 )
 
 type cmdType int