+++ client/v2/autocli/flag/dec_coin.go
@@ -8,7 +8,8 @@ import (
 	"google.golang.org/protobuf/reflect/protoreflect"
 
 	basev1beta1 "cosmossdk.io/api/cosmos/base/v1beta1"
-	"cosmossdk.io/client/v2/internal/coins"
+
+	"github.com/cosmos/cosmos-sdk/client/v2/internal/coins"
 )
 
 type decCoinType struct{}