+++ client/v2/autocli/flag/map.go
@@ -9,7 +9,7 @@ import (
 	"github.com/spf13/pflag"
 	"google.golang.org/protobuf/reflect/protoreflect"
 
-	"cosmossdk.io/client/v2/autocli/flag/maps"
+	"github.com/cosmos/cosmos-sdk/client/v2/autocli/flag/maps"
 )
 
 func bindSimpleMapFlag(flagSet *pflag.FlagSet, keyKind, valueKind protoreflect.Kind, name, shorthand, usage string) HasValue {