+++ client/v2/autocli/util.go
@@ -7,7 +7,7 @@ import (
 	"github.com/spf13/cobra"
 	"google.golang.org/protobuf/reflect/protoreflect"
 
-	"cosmossdk.io/client/v2/internal/strcase"
+	"github.com/cosmos/cosmos-sdk/client/v2/internal/strcase"
 )
 
 // findSubCommand finds a sub-command of the provided command whose Use