+++ server/grpc/gogoreflection/fix_registration.go
@@ -29,6 +29,7 @@ func getMessageType(name string) reflect.Type {
 	return proto.MessageType(name) //nolint:staticcheck // keep for backward compatibility
 }
 
+//nolint:staticcheck // keep for backward compatibility
 func getExtension(extID int32, m proto.Message) *gogoproto.ExtensionDesc {
 	// check first in gogoproto registry
 	for id, desc := range gogoproto.RegisteredExtensions(m) {
@@ -38,7 +39,7 @@ func getExtension(extID int32, m proto.Message) *gogoproto.ExtensionDesc {
 	}
 
 	// check into proto registry
-	for id, desc := range proto.RegisteredExtensions(m) { //nolint:staticcheck // keep for backward compatibility
+	for id, desc := range proto.RegisteredExtensions(m) {
 		if id == extID {
 			return &gogoproto.ExtensionDesc{
 				ExtendedType:  desc.ExtendedType,