+++ tests/e2e/bank/cli_test.go
@@ -7,9 +7,10 @@ import (
 	"testing"
 
 	"cosmossdk.io/simapp"
-	"github.com/cosmos/cosmos-sdk/testutil/network"
 
 	"github.com/stretchr/testify/suite"
+
+	"github.com/cosmos/cosmos-sdk/testutil/network"
 )
 
 func TestE2ETestSuite(t *testing.T) {